requirements.txt
# Основные runtime-зависимости
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.5.0
PyYAML>=6.0.0

# RAG-поиск
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Тесты
pytest>=7.4.0
